requirements.txt
Pillow==8.2.0
ipython==7.22.0